+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 ### Fixed
 - disabled test parameter
 
+### Improved
+- added the ability to specify an information message for the courier who arrives to pickup the shipments
+- added the ability to specify the arrival date and time when calling a courier
+
+### Changed
+- updated the courier invitation function according to changes in the API library
+
+### Updated
+- itella-api library to v2.4.3
+
 ## [1.5.5] - 2025-04-03
 ### Fixed
 - fixed loading of external scripts on the Checkout page so that they are loaded only on this page